requirements.txt
# ============================================================================
# Kimi K2.5 Agent Swarm - Production Dependencies
# NO MOCK DEPENDENCIES - All real integrations
# ============================================================================

# Core FastAPI and async framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# HTTP clients for real API calls
httpx>=0.27.0
aiohttp==3.9.1
requests==2.31.0

# Database - PostgreSQL with vector support (NO MOCKS)
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis for caching and pub/sub
redis==5.0.1
aioredis==2.0.1

# Vector databases (REAL implementations)
chromadb==0.4.22
qdrant-client==1.7.0

# Real embeddings - OpenAI, Cohere (NO HASH FAKES)
openai==1.10.0
cohere==4.39
anthropic==0.8.1

# LLM providers (REAL APIs)
together==0.2.11

# Data processing
numpy==1.26.3
pandas==2.1.4

# Security (production-grade)
bcrypt==4.1.2
argon2-cffi==23.1.0
cryptography==42.0.0
pyjwt==2.8.0

# Testing (integration tests with real APIs)
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov==4.1.0

# Monitoring and observability
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# Logging
structlog==24.1.0

# Rate limiting
slowapi==0.1.9

# CORS and security headers
python-multipart==0.0.6

# Task queue for background jobs
celery==5.3.4

# Optional enhanced features
aiofiles>=23.2.1
rich>=13.7.0

# Development tools
black>=24.0.0
ruff>=0.1.0
